remove legacy FileVersionInfo.format_ls_entry and FileVersionInfo.format_folder_ls_entry
