+++ b2sdk/v1/file_version.py
@@ -1,0 +1,36 @@
+######################################################################
+#
+# File: b2sdk/v1/file_version.py
+#
+# Copyright 2021 Backblaze Inc. All Rights Reserved.
+#
+# License https://www.backblaze.com/using_b2_code.html
+#
+######################################################################
+
+import datetime
+
+from b2sdk import _v2 as v2
+
+
+# override to retain old formatting methods
+class FileVersionInfo(v2.FileVersionInfo):
+    LS_ENTRY_TEMPLATE = '%83s  %6s  %10s  %8s  %9d  %s'  # order is file_id, action, date, time, size, name
+
+    def format_ls_entry(self):
+        dt = datetime.datetime.utcfromtimestamp(self.upload_timestamp / 1000)
+        date_str = dt.strftime('%Y-%m-%d')
+        time_str = dt.strftime('%H:%M:%S')
+        size = self.size or 0  # required if self.action == 'hide'
+        return self.LS_ENTRY_TEMPLATE % (
+            self.id_,
+            self.action,
+            date_str,
+            time_str,
+            size,
+            self.file_name,
+        )
+
+    @classmethod
+    def format_folder_ls_entry(cls, name):
+        return cls.LS_ENTRY_TEMPLATE % ('-', '-', '-', '-', 0, name)
